+++ src/llama_inference/__init__.py
@@ -1,5 +1,5 @@
 """Inference API for LLaMA"""
 
-from .api import LLaMAInference
+from .model import LLaMAInference
 
 __version__ = "0.0.0"
